- I was wrong. Two commits ago about wxTheApp->ExitMainLoop() while it works it is not the proper way to force exit of wxApp. It is fixed. The issue I believe was due to double deletion of the DialogBox. According to this article:
https://wiki.wxwidgets.org/Avoiding_Memory_Leaks#Automatic_leak_detection_with_Microsoft_VisualC.2B.2B

Modal dialogs are safe to delete manually using "delete". But because modal
dialogs have its own event handling that differs than modeless frames. Modal
dialogs can be created on stack (as normal modeless frames) BUT should not be
Destroy()ed manually. Otherwise, double deletion and crash.

While debugging, it was giving acess violation for the delete operator, and
if I got around that, it waits for the object to respond "thread" causing
the Emu to be running in limbo until closed via task manager.

- Anyways, changed the code a bit while testing for the proper solution to the
problem above. Notably the TwinPad_Frame from global to local being shown and
destroyed after being closed. Releasing its memory so consequent calls will
create TwinPad_Frame again. Much cleaner.

Note: DON'T use EndModal(retValue) it causes bad behavior. Instead, Hide() then
Close() the TwinPad_Frame, which gets deleted after that. That way the Emu will
not be hidden/shown causing flicker like effect.

Note2: I finally found a proper DLL example:

https://forums.wxwidgets.org/viewtopic.php?p=5964#5964

But unfortunately, creating a dummy wxWindow and assign the HWND to it and make
it as a parent for the dialog didn't work for me. I will do further testing
later on this.

+++ TwinPad/source/twinpad/plugin/headers/Externals.h
@@ -24,10 +24,10 @@ typedef struct {
 	u8 button;
 } Analog;
 
-extern unsigned short mousePAD;
-extern unsigned short mouseSensitivity;
-extern unsigned short mouseScrollUp;
-extern unsigned short mouseScrollDown;
+extern unsigned int mousePAD;
+extern unsigned int mouseSensitivity;
+extern unsigned int mouseScrollUp;
+extern unsigned int mouseScrollDown;
 extern char MouseButtonMap[8];
 extern bool inside;  // if mouse inside the window..
 extern Analog lanalog[2];
